\name{obs.agree-package}
\alias{obs.agree-package}
\alias{obs.agree}
\docType{package}
\title{
Observer agreement package 
}
\description{

The package includes two functions for measuring agreement. Raw Agreement Indices to categorical data and Information-Based Measure of Disagreement to continuous data. It can be used for multiple raters and multiple readings cases.
}
\details{
\tabular{ll}{
Package: \tab obs.agree\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-09\cr
License: \tab GPL-2 \cr
}

}
\author{Teresa Henriques, Luis Antunes and Cristina Costa-Santos }

\keyword{ package }

